importSystemStructure: Separated API and CLI code
