+++ docs/source/cli.rst
@@ -5,5 +5,5 @@ CLI Documentation
    :maxdepth: 2
 
    cli.ospCaseBuilder
-   cli.watchCosim
    cli.importSystemStructure
+   cli.watchCosim
