+++ setup.cfg
@@ -53,7 +53,7 @@ exclude =
 console_scripts =
     ospCaseBuilder = ospx.cli.ospCaseBuilder:main
     watchCosim = ospx.watch.cli.watchCosim:main
-    importSystemStructure = ospx.importer.cli.importSystemStructure:main
+    importSystemStructure = ospx.cli.importSystemStructure:main
 
 [flake8]
 ignore = E116, E117, E123, E126, E226, E133, E251, E501, E701, W503
